FM-189: extract announcement list (not ready)
