+++ constants/locales.js
@@ -64,4 +64,6 @@ export default {
   COAT_COLORS: 'Umaszczenie',
   FINISH: 'Zakończ',
   ARCHIVE: 'Archiwizuj',
+  YOU_DONT_HAVE_ANY_ANNOUNCEMENTS_YET: 'Nie masz jeszcze żadnych ogłoszeń',
+  USER_DONT_HAVE_ANY_ANNOUNCEMENTS_YET: 'Użytkownik nie ma jeszcze żadnych ogłoszeń',
 };
